# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Restaurant Documentation",
        "description": "L5 Swagger OpenApi description for restaurant project",
        "contact": {
            "email": "drrifatalashwad0@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/v1.0/auth/register-with-garage": {
            "post": {
                "tags": [
                    "garage_management"
                ],
                "summary": "This method is to store user with garage",
                "description": "This method is to store user with garage",
                "operationId": "registerUserWithGarage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user",
                                    "garage",
                                    "service"
                                ],
                                "properties": {
                                    "user": {
                                        "type": "string",
                                        "format": "array",
                                        "example": {
                                            "first_Name": "Rifat",
                                            "last_Name": "Al-Ashwad",
                                            "email": "rifatalashwad@gmail.com",
                                            "password": "12345678",
                                            "password_confirmation": "12345678",
                                            "phone": "01771034383",
                                            "image": "https://images.unsplash.com/photo-1671410714831-969877d103b1?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=387&q=80",
                                            "address_line_1": "Dhaka",
                                            "address_line_2": "Dinajpur",
                                            "country": "Bangladesh",
                                            "city": "Dhaka",
                                            "postcode": "Dinajpur"
                                        }
                                    },
                                    "garage": {
                                        "type": "string",
                                        "format": "array",
                                        "example": {
                                            "name": "ABCD Garage",
                                            "about": "Best Garage in Dhaka",
                                            "web_page": "https://www.facebook.com/",
                                            "phone": "01771034383",
                                            "email": "rifatalashwad@gmail.com",
                                            "additional_information": "No Additional Information",
                                            "address_line_1": "Dhaka",
                                            "address_line_2": "Dinajpur",
                                            "country": "Bangladesh",
                                            "city": "Dhaka",
                                            "postcode": "Dinajpur",
                                            "logo": "https://images.unsplash.com/photo-1671410714831-969877d103b1?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=387&q=80",
                                            "is_mobile_garage": true,
                                            "wifi_available": true,
                                            "labour_rate": 500
                                        }
                                    },
                                    "service": {
                                        "type": "string",
                                        "format": "array",
                                        "example": [
                                            {
                                                "automobile_category_id": 1,
                                                "services": [
                                                    {
                                                        "id": 1,
                                                        "checked": true,
                                                        "sub_services": [
                                                            {
                                                                "id": 1,
                                                                "checked": true
                                                            },
                                                            {
                                                                "id": 2,
                                                                "checked": false
                                                            }
                                                        ]
                                                    }
                                                ],
                                                "automobile_makes": [
                                                    {
                                                        "id": 1,
                                                        "checked": true,
                                                        "models": [
                                                            {
                                                                "id": 1,
                                                                "checked": true
                                                            },
                                                            {
                                                                "id": 2,
                                                                "checked": false
                                                            }
                                                        ]
                                                    }
                                                ]
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocesseble Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1.0/garages/{perPage}": {
            "get": {
                "tags": [
                    "garage_management"
                ],
                "summary": "This method is to get garages",
                "description": "This method is to get garages",
                "operationId": "getGarages",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "path",
                        "description": "perPage",
                        "required": true,
                        "example": "6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocesseble Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}